% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_imputer.R
\name{date_imputer}
\alias{date_imputer}
\title{Imputation function for a date column}
\usage{
date_imputer(x, col, method)
}
\arguments{
\item{x}{input data tibble}

\item{col}{name of the column to modify}

\item{method}{name of the imputing strategy. They include \link{'rand', 'frequent', and 'median'}.}
}
\value{
output data tibble after imputation
}
\description{
Imputation function for a date column
}
\examples{
input_df <- data.frame(a = (c(as.Date('1990/02/17'), as.Date('1991/02/17'), NaN, as.Date('1992/02/17'))))
date_imputer(input_df, "a", "median")
}
