% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_imputer.R
\name{num_imputer}
\alias{num_imputer}
\title{Imputation function for a numerical column}
\usage{
num_imputer(x, col)
}
\arguments{
\item{x}{input data tibble}

\item{col}{name of the column to modify and that should be a string or charactor}
}
\value{
output data tibble after imputation
}
\description{
Imputation function for a numerical column
}
\examples{
input_df <- data.frame(a = c(1,2,3,4,5,6, NA))
num_imputer(input_df, "a")
}
