% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_imputer.R
\name{num_imputer}
\alias{num_imputer}
\title{Imputation function for a numerical column}
\usage{
num_imputer(x, col, method)
}
\arguments{
\item{x}{input data tibble}

\item{col}{name of the column to modify}

\item{method}{name of the imputing strategy. They include \link{'knn','quant_rand','mean','median','mode','bayes'}, the default method is knn.}
}
\value{
output data tibble after imputation
}
\description{
Imputation function for a numerical column
}
\examples{
num_imputer(input_tibble, col_name, "mean")
}
